opencv/requirements.txt
aiohttp
nats-py
numpy
